e2e_tests/requirements.txt
selenium==4.18.1
pytest==7.4.3
pytest-html==4.1.1
webdriver-manager==4.0.1
python-dotenv==1.0.0 